# demo/src/main/resources/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Restcountries",
    "description": "API for restcountries.eu",
    "version": "1.0.0"
  },
  "host": "restcountries.eu",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/rest/v2/alpha/{alphacode}": {
      "get": {
        "tags": [
          "Rest"
        ],
        "description": "Get By Alphacode",
        "operationId": "GetByAlphacode",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "alphacode",
            "in": "path",
            "description": "alphacode",
            "required": true,
            "type": "string",
            "x-example": "ZA"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetAll"
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "definitions": {
    "GetAll": {
      "description": "Model for GetAll",
      "required": [
        "alpha2Code"
      ],
      "properties": {
        "postalCode": {
            "$ref": "#/definitions/PostalCode"
        },
        "name": {
          "$ref": "#/definitions/Name"
        }
      }
    },
...
    "Name": {
      "description": "Model for Name",
      "properties": {
        "common": {
          "type": "string"
        }
      }
    },
    "PostalCode": {
      "description": "Model for PostalCode",
      "properties": {
        "format": {
          "type": "string"
        },
        "regex": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {},
  "security": [],
...
}